283. 移动零

+++ leetcode/src/main/java/com/liufei/top100/Top036_middle.java
@@ -82,7 +82,7 @@ public class Top036_middle {
             for (int j = 0; j < 9; j++) {
                 char num = board[i][j];
                 if (num != '.') {
-                    int n = (int)num;
+                    int n = num;
                     int box_index = (i / 3 ) * 3 + j / 3;
 
                     // keep the current cell value
